🔨 Add: CodeEditor 컴포넌트 추가
